+++ core/menu.sh
@@ -22,11 +22,7 @@ source "$ROOT_DIR/modules/system_cleanup.sh"
 # shellcheck disable=SC1091
 source "$ROOT_DIR/modules/scripts_hub.sh"
 # shellcheck disable=SC1091
-source "$ROOT_DIR/modules/docker_manager.sh"
-# shellcheck disable=SC1091
-source "$ROOT_DIR/modules/singbox.sh"
-# shellcheck disable=SC1091
-source "$ROOT_DIR/modules/menus/load.sh"
+source "$ROOT_DIR/modules/kejilion_bridge.sh"
 run_action() {
   local action_name="$1"
   shift
@@ -65,9 +61,6 @@ main_menu() {
     menu_item "13" "$(msg menu_label_13)"
     menu_item "14" "$(msg menu_label_14)"
     menu_item "15" "$(msg menu_label_15)"
-    menu_item "16" "$(msg menu_label_16)"
-    menu_item "17" "$(msg menu_label_17)"
-    menu_item "18" "$(msg menu_label_18)"
     echo "----------------------------------------"
     menu_item "99" "$(msg menu_label_99)"
     menu_item "88" "$(msg menu_label_88)"
@@ -95,60 +88,48 @@ main_menu() {
         run_action "scripts_hub" scripts_hub
         ;;
       5)
-        log_action "menu:docker_manager"
-        run_action "docker_manager" docker_manager
+        log_action "menu:basic_tools"
+        run_action "basic_tools_menu" basic_tools_menu
         ;;
       6)
-        log_action "menu:network_accel"
-        run_action "network_accel_menu" network_accel_menu
+        log_action "menu:bbr_management"
+        run_action "bbr_management_menu" bbr_management_menu
         ;;
       7)
-        log_action "menu:network_test"
-        run_action "network_test_menu" network_test_menu
+        log_action "menu:docker_management"
+        run_action "docker_management_menu" docker_management_menu
         ;;
       8)
-        log_action "menu:security"
-        run_action "security_menu" security_menu
+        log_action "menu:warp_management"
+        run_action "warp_management_menu" warp_management_menu
         ;;
       9)
-        log_action "menu:ldnmp"
-        run_action "ldnmp_menu" ldnmp_menu
+        log_action "menu:network_test_suite"
+        run_action "network_test_suite_menu" network_test_suite_menu
         ;;
       10)
-        log_action "menu:app_market"
-        run_action "app_market_menu" app_market_menu
+        log_action "menu:oracle_cloud_suite"
+        run_action "oracle_cloud_suite_menu" oracle_cloud_suite_menu
         ;;
       11)
-        log_action "menu:workspace"
-        run_action "workspace_menu" workspace_menu
+        log_action "menu:ldnmp_site_suite"
+        run_action "ldnmp_site_suite_menu" ldnmp_site_suite_menu
         ;;
       12)
-        log_action "menu:system_tools"
-        run_action "system_tools_menu" system_tools_menu
+        log_action "menu:app_marketplace"
+        run_action "app_marketplace_menu" app_marketplace_menu
         ;;
       13)
-        log_action "menu:backup"
-        run_action "backup_menu" backup_menu
+        log_action "menu:workspace_suite"
+        run_action "workspace_suite_menu" workspace_suite_menu
         ;;
       14)
-        log_action "menu:cron_center"
-        run_action "cron_center_menu" cron_center_menu
+        log_action "menu:system_tools_suite"
+        run_action "system_tools_suite_menu" system_tools_suite_menu
         ;;
       15)
-        log_action "menu:cluster"
-        run_action "cluster_menu" cluster_menu
-        ;;
-      16)
-        log_action "menu:oracle_cloud"
-        run_action "oracle_cloud_menu" oracle_cloud_menu
-        ;;
-      17)
-        log_action "menu:game_server"
-        run_action "game_server_menu" game_server_menu
-        ;;
-      18)
-        log_action "menu:ai_workspace"
-        run_action "ai_workspace_menu" ai_workspace_menu
+        log_action "menu:cluster_control_suite"
+        run_action "cluster_control_suite_menu" cluster_control_suite_menu
         ;;
       99|00)
         log_action "menu:self_update"